+++ src/js/components/common/icons/ExpandIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-export const ExpandIcon = () =>
-	<svg width="18" height="18" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg">
+export const ExpandIcon = (props: React.SVGProps<SVGSVGElement>) =>
+	<svg width="18" height="18" viewBox="0 0 18 18" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
 		<path d="M12 3H15V6" />
 		<path d="M10.5 7.5L15 3L10.5 7.5Z" />
 		<path d="M6 15H3V12" />